feat: CHASE-55 users presentationレイヤーをdataSourceパターンに再構成
- routes/profile/とroutes/settings/の個別モジュール作成
- shared/フォルダでcommon-schemasとerror-handling統合
- dependency injectionパターンでサービス統合
- testClientからapp.request()パターンに移行
- エラーハンドリングロジック改善（内部エラーは500で返却）
- 全テスト219件が正常通過
